+++ src/main/java/org/betonquest/betonquest/compatibility/luckperms/LuckPermsIntegrator.java
@@ -2,35 +2,45 @@ package org.betonquest.betonquest.compatibility.luckperms;
 
 import net.luckperms.api.LuckPerms;
 import net.luckperms.api.context.ContextCalculator;
+import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.compatibility.Integrator;
+import org.betonquest.betonquest.compatibility.luckperms.permission.LuckPermsEventFactory;
 import org.betonquest.betonquest.exceptions.HookException;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.RegisteredServiceProvider;
 
-@SuppressWarnings({"PMD.CommentRequired", "PMD.AtLeastOneConstructor"})
+@SuppressWarnings("PMD.CommentRequired")
 public class LuckPermsIntegrator implements Integrator {
-    private LuckPerms api;
+
+    private final BetonQuest instance;
+
+    private LuckPerms luckPermsAPI;
 
     private ContextCalculator<Player> tagCalculator;
+
+    public LuckPermsIntegrator() {
+        instance = BetonQuest.getInstance();
+    }
 
     @Override
     public void hook() throws HookException {
         final RegisteredServiceProvider<LuckPerms> provider = Bukkit.getServicesManager().getRegistration(LuckPerms.class);
         if (provider != null) {
-            api = provider.getProvider();
+            luckPermsAPI = provider.getProvider();
             tagCalculator = TagCalculatorUtils.getTagContextCalculator();
-            api.getContextManager().registerCalculator(tagCalculator);
+            luckPermsAPI.getContextManager().registerCalculator(tagCalculator);
+            instance.registerNonStaticEvent("luckperms", new LuckPermsEventFactory(luckPermsAPI));
         }
     }
 
     @Override
-    @SuppressWarnings("PMD.UncommentedEmptyMethodBody")
     public void reload() {
+        // Empty
     }
 
     @Override
     public void close() {
-        api.getContextManager().unregisterCalculator(tagCalculator);
+        luckPermsAPI.getContextManager().unregisterCalculator(tagCalculator);
     }
 }