+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -193,7 +193,6 @@ import org.betonquest.betonquest.objectives.MobKillObjective;
 import org.betonquest.betonquest.objectives.PasswordObjective;
 import org.betonquest.betonquest.objectives.PickupObjective;
 import org.betonquest.betonquest.objectives.ResourcePackObjective;
-import org.betonquest.betonquest.objectives.RespawnObjective;
 import org.betonquest.betonquest.objectives.RideObjective;
 import org.betonquest.betonquest.objectives.ShearObjective;
 import org.betonquest.betonquest.objectives.SmeltingObjective;
@@ -981,7 +980,6 @@ public class BetonQuest extends JavaPlugin {
         registerObjectives("variable", VariableObjective.class);
         registerObjectives("kill", KillPlayerObjective.class);
         registerObjectives("interact", EntityInteractObjective.class);
-        registerObjectives("respawn", RespawnObjective.class);
         registerObjectives("breed", BreedObjective.class);
         registerObjectives("command", CommandObjective.class);
         if (PaperLib.isPaper()) {