+++ CHANGELOG.md
@@ -15,7 +15,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `aureliumstatslevel` condition was renamed to `auraskillsstatslevel`
   - `aureliumskillsxp` event was renamed to `auraskillsxp`
 - prevent reply when the text is not completely displayed for the SlowTellRaw conversation IO
-- `resourcepack` objective is now paper only 
+- `resourcepack` objective is now paper only
+- `die` objective now support respawns without canceling the actual death of the player
 ### Deprecated
 ### Removed
 ### Fixed