+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - prevent reply when the text is not completely displayed for the SlowTellRaw conversation IO
 - `resourcepack` objective is now paper only
 - `die` objective now support respawns without canceling the actual death of the player
+- `ph` variable can now be used static
 ### Deprecated
 ### Removed
 ### Fixed