+++ CHANGELOG.md
@@ -26,6 +26,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - head `QuestItem` matches even if they are distinct 
 - modulo operator in math variable can now be used by escaping the percent sign
 - `weather` event not static callable even when providing a world
+- hologram checks the conditions if the player is outside the `max_range`.
 ### Security
 
 ## [2.1.1] - 2024-05-09